+++ src/extension/src/Cleaning/generelClean.js
@@ -65,7 +65,6 @@ function correctImageFormat(imgScr) {
 /*Final cleanup function */
 export function generalClean(readabilityContent) {
   let cleanContent = removeSVG(readabilityContent);
-  cleanContent = removeLinks(cleanContent);
   cleanContent = removeFigures(cleanContent);
   return cleanContent;
 }
@@ -78,19 +77,6 @@ function removeSVG(readabilityContent) {
   return div.innerHTML;
 }
 
-function removeLinks(readabilityContent) {
-  const div = document.createElement("div");
-  div.innerHTML = readabilityContent;
-  var links = div.getElementsByTagName("a");
-  while (links.length) {
-    var parent = links[0].parentNode;
-    while (links[0].firstChild) {
-      parent.insertBefore(links[0].firstChild, links[0]);
-    }
-    parent.removeChild(links[0]);
-  }
-  return div.innerHTML;
-}
 
 function removeFigures(readabilityContent) {
   const div = document.createElement("div");