+++ src/extension/src/popup/Popup.js
@@ -10,7 +10,7 @@ import { HeadingContainer, PopUp, BottomContainer } from "./Popup.styles";
 import PopupContent from "./PopupContent";
 import { EXTENSION_SOURCE } from "../constants";
 import { checkLanguageSupport, setUserInLocalStorage } from "./functions";
-import { StyledPrimaryButton } from "../Modal/Buttons.styles";
+import { StyledPrimaryButton } from "../InjectedReaderApp/Buttons.styles";
 import { API_URL, WEB_URL } from "../../../config";
 import { BROWSER_API } from "../utils/browserApi";
 
@@ -25,6 +25,9 @@ export default function Popup({ loggedIn }) {
   const [tab, setTab] = useState();
   const [isReadable, setIsReadable] = useState();
   const [languageSupported, setLanguageSupported] = useState();
+  const [articleData, setArticleData] = useState();
+  const [fragmentData, setFragmentData] = useState();
+  const [loadingProgress, setLoadingProgress] = useState("Analyzing page...");
 
   useEffect(() => {
     if (loggedIn) {
@@ -66,7 +69,7 @@ export default function Popup({ loggedIn }) {
         setIsReadable(true);
         api.session = user.session;
         if (api.session !== undefined) {
-          checkLanguageSupport(api, tab, setLanguageSupported);
+          checkLanguageSupport(api, tab, setLanguageSupported, setArticleData, setLoadingProgress, setFragmentData);
         }
       }
     }
@@ -102,6 +105,9 @@ export default function Popup({ loggedIn }) {
             tab={tab}
             api={api}
             sessionId={user.session}
+            articleData={articleData}
+            fragmentData={fragmentData}
+            loadingProgress={loadingProgress}
           ></PopupContent>
         </PopUp>
       );